added email support for user notificatios

+++ Backend/src/controllers/authController.js
@@ -125,7 +125,7 @@ exports.login = catchAsync(async (req, res, next) => {
   }
 
   // 2. Check if account exists && password is correct
-  const user = await User.findOne({ email }).select("password");
+  const user = await User.findOne({ email }).select("+password");
 
   if (!user || !(await user.correctPassword(password, user.password))) {
     return next(new AppError("Incorrect email or password.", 400));
@@ -316,11 +316,16 @@ exports.updateProfile = catchAsync(async (req, res, next) => {
 
 exports.deleteAccount = catchAsync(async (req, res, next) => {
   const userId = req.user.id;
+  const { password } = req.body;
 
-  const user = await User.findById(userId);
+  const user = await User.findById(userId).select("+password");
 
   if (!user) {
     return next(new AppError("User not found.", 404));
+  }
+
+  if (!(await user.correctPassword(password, user.password))) {
+    return next(new AppError("Incorrect password.", 400));
   }
 
   await User.findByIdAndDelete(user._id);
